pyproject.toml
[project]
name = "waffle-maker"
version = "0.1.0"
description = "ROS 2 Waffle Dashboard — static site generator for PR triage"
requires-python = ">=3.11"
dependencies = [
    "PyGithub>=2.0",
    "Jinja2>=3.0",
    "PyYAML>=6.0",
    "requests>=2.28",
]

[project.scripts]
waffle-maker = "generate_site:main"
